move:#42 PATCH 요청 코드 옮기기
